+++ CHANGELOG.md
@@ -16,6 +16,7 @@ None
 - Fix the transfer maps in `Drift` and `Dipole`; Add R56 in horizontal and vertical correctors modelling (see #90) (@cr-xu)
 - Fix fringe_field_exit of `Dipole` is overwritten by `fringe_field` bug (see #99) (@cr-xu)
 - Fix error caused by mismatched devices on machines with CUDA GPUs (see #97) (@jank324)
+- Fix error raised when tracking a `ParameterBeam` through an active `BPM` (see #101) (@jank324)
 
 ### 🐆 Other